Refactor MCPServer form stylings
