+++ workspaces/mi/mi-visualizer/src/views/Forms/MCPServerForm/styles.ts
@@ -18,31 +18,6 @@
 
 import styled from '@emotion/styled';
 
-export const DialogField = styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    margin-bottom: 15px;
-`;
-
-export const DialogButtonGroup = styled.div`
-    display: flex;
-    gap: 10px;
-    justify-content: space-between;
-    margin-top: 15px;
-    padding-top: 15px;
-    border-top: 1px solid var(--vscode-panel-border);
-`;
-
-export const DialogTitle = styled.div`
-    text-align: center;
-    margin-bottom: 20px;
-    font-size: 16px;
-    font-weight: 600;
-    color: var(--vscode-editor-foreground);
-`;
-
-// Common list components
 export const ItemsList = styled.div`
     background: var(--vscode-input-background);
     border: 1px solid var(--vscode-input-border);
@@ -85,19 +60,6 @@ export const SelectAllRow = styled.div`
     background: var(--vscode-list-activeSelectionBackground);
     cursor: pointer;
     user-select: none;
-`;
-
-// Common layout components
-export const FlexRow = styled.div`
-    display: flex;
-    align-items: center;
-    gap: 6px;
-`;
-
-export const FlexRowStart = styled.div`
-    display: flex;
-    align-items: flex-start;
-    gap: 8px;
 `;
 
 export const CustomInputsContainer = styled.div`
